Reorganized gameClasses folder
