+++ src/game/gameClasses/CMakeLists.txt
@@ -1,3 +1,1 @@
-target_sources(${PROJECT_NAME} PRIVATE
-        camera.c)
-
+add_subdirectory(camera)
