+++ src/game/gameClasses/camera/CMakeLists.txt
@@ -1,0 +1,2 @@
+target_sources(${PROJECT_NAME} PRIVATE
+        camera.c)
